Rename drld_parser to metis_drld
